+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds a "Unique Players per Channel" chart to the analytics page.
+
 ## [v18.9.3](https://github.com/lexicalunit/spellbot/releases/tag/v18.9.3) - 2026-05-14
 
 ### Changed